requirements.txt
pyannote.audio
torchvision
openai-whisper
